+++ src/common/Smi.Common/Execution/MicroserviceHost.cs
@@ -1,10 +1,5 @@
 
-using System;
-using System.Diagnostics;
-using System.IO;
-using System.Reflection;
 using DicomTypeTranslation;
-using DicomTypeTranslation.Helpers;
 using JetBrains.Annotations;
 using NLog;
 using RabbitMQ.Client;
@@ -13,6 +8,10 @@ using Smi.Common.Helpers;
 using Smi.Common.Messages;
 using Smi.Common.Messaging;
 using Smi.Common.Options;
+using System;
+using System.Diagnostics;
+using System.IO;
+using System.Reflection;
 
 namespace Smi.Common.Execution
 {
@@ -205,8 +204,6 @@ namespace Smi.Common.Execution
         /// <param name="exception"></param>
         public void Fatal(string msg, Exception exception)
         {
-            Logger.Fatal(exception, msg);
-
             if (_stopCalled)
                 return;