+++ src/microservices/Microservices.CohortExtractor/Execution/RequestFulfillers/FakeFulfiller.cs
@@ -7,6 +7,10 @@ using System.Text.RegularExpressions;
 
 namespace Microservices.CohortExtractor.Execution.RequestFulfillers
 {
+    /// <summary>
+    /// Fake <see cref="IExtractionRequestFulfiller"/> that automatically finds all UIDs that you ask it to look up and returns
+    /// a single image name for each requested UID.  The filename will be the UID(s) you asked for
+    /// </summary>
     public class FakeFulfiller : IExtractionRequestFulfiller
     {
         protected readonly Logger Logger;
@@ -17,8 +21,6 @@ namespace Microservices.CohortExtractor.Execution.RequestFulfillers
         public FakeFulfiller()
         {
             Logger = LogManager.GetCurrentClassLogger();
-
-            Logger.Debug("Faking a filename");
         }
 
         public IEnumerable<ExtractImageCollection> GetAllMatchingFiles(ExtractionRequestMessage message, IAuditExtractions auditor)